+++ CHANGELOG.md
@@ -38,6 +38,7 @@ No methods are defined to require OAuth yet!
 - After finished solr indexing the `/solr/index`-endpoint now updates the index for the suggest component as well.
 - Added the ability to enrich metadata with extra fragments from a seperate webservice when harvesting records from preservica
 - handle errors when calling fragment service
+- Added manual batchjob to generate Kaltura XML for bulk uploads.
 
 ### Changed  
 - Changed when DsPreservicaClient is initialized.